+++ CHANGES.md
@@ -14,6 +14,8 @@ Changes in CUPS v2.4.2 (TBA)
   IPP Everywhere queue (Issues #340, #343)
 - Re-added LibreSSL/OpenSSL support (Issue #362)
 - Updated the Solaris smf service file (Issue #368)
+- The scheduler now regenerates the PPD cache information after changing the
+  "cupsd.conf" file (Issue #371)
 - Updated the scheduler to set "auth-info-required" to "username,password" if a
   backend reports it needs authentication info but doesn't set a method for
   authentication (Issue #373)